Voice Ai Core Backend/tts/human_tts/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch>=2.1.0
torchaudio>=2.1.0
transformers>=4.40.0
librosa==0.10.2
audiomentations==0.36.0
soundfile==0.12.1
numpy==1.26.4
scipy==1.12.0
python-dotenv==1.0.1
aiofiles==23.2.1
accelerate>=0.26.0
parler-tts
